+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Add a `--cointracking` flag to format the data for use with CoinTracking's
+  Bulk Import feature.
 * Add an `--output-file` CLI argument to make the executable write to a file
   instead of print to stdout.
 * Use the `cmdargs` package to parse CLI arguments.